Fix bug where rasterize() hangs on Apple silicon
If any edges of the triangle are horizontal or vertical, it's possible
for the raXX variables, and thus d, to get a division by zero leading
to +Inf. On Intel, when +Inf is converted to int and back with
  d = float64(int(d))
the result is a very large negative number which the existing check
converted to zero. On ARM (or at least on Apple Silicon which is where I
ran into the problem) the result is a very large positive number,
causing the function to hang.